imporved code, ui
