+++ src/__test__/04.test.js
@@ -86,6 +86,9 @@ describe('unit test link', () => {
       fireEvent.click(screen.getAllByRole('button', { name: 'Cart' })[1])
     })
     await waitFor(() => {
+       fireEvent.click(screen.getByRole('tab', { name: 'Cart', selected: false }))
+    })
+    await waitFor(() => {
       fireEvent.click(screen.getByRole('button', { name: 'Buy' }))
     })
     const textMail = screen.getByRole('textbox', { name: 'Email' })
