+++ src/__test__/common.js
@@ -10,8 +10,11 @@ import React from 'react'
 import {render} from '@testing-library/react'
 import {Store,Order,Cart} from '../App'
 import {Signin} from '../components/Signin'
-import {ShopContextProvider, CartContextProvider} from '../store'
+import {StoreContextProvider} from '../store'
 import {BrowserRouter,Routes,Route} from 'react-router-dom'
+import {Shop} from '../components/Shop'
+import {Header} from '../components/Header'
+import {Cart as CartPart} from '../components/Cart'
 
 export const response = class {
   constructor(filename) {
@@ -24,29 +27,26 @@ export const response = class {
 }
 export const testRender = () => {
   return render(
-    <ShopContextProvider>
-      <CartContextProvider>
-        <BrowserRouter>
-          <Store />
-        </BrowserRouter>
-      </CartContextProvider>
-    </ShopContextProvider>
+    <StoreContextProvider>
+      <BrowserRouter>
+        <Shop />
+        <CartPart />
+      </BrowserRouter>
+    </StoreContextProvider>
   )
 }
 export const testLoginRender = () => {
   return render(
-    <ShopContextProvider>
-      <CartContextProvider>
-        <BrowserRouter>
-          <Routes>
-            <Route path={'/'} element={<Signin />} />
-            <Route path={'/shop'} element={<Store />} />
-            <Route path={'/cart'} element={<Cart />} />
-            <Route path={'/order'} element={<Order />} />
-            <Route path={'*'} element={<Signin />} />
-          </Routes>
-        </BrowserRouter>
-      </CartContextProvider>
-    </ShopContextProvider>
+    <StoreContextProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path={'/'} element={<Signin />} />
+          <Route path={'/shop'} element={<Store />} />
+          <Route path={'/cart'} element={<Cart />} />
+          <Route path={'/order'} element={<Order />} />
+          <Route path={'*'} element={<Signin />} />
+        </Routes>
+      </BrowserRouter>
+    </StoreContextProvider>
   )
 }
