+++ src/actions/shopAction.js
@@ -6,12 +6,8 @@
  */
 'use strict'
 
-import {CHANGE_ITEM_LIST, SIGNIN_USERNAME, ADD_ORDER} from '../reducers/shopReducer'
+import {SIGNIN_USERNAME, ADD_ORDER} from '../reducers/shopReducer'
 
-export const changeItemList = (jsonfile) => ({
-  type: CHANGE_ITEM_LIST,
-  payload: {jsonfile},
-})
 export const signinUsername = (username) => ({
   type: SIGNIN_USERNAME,
   payload: {username},
