+++ src/reducers/shopReducer.js
@@ -6,7 +6,6 @@
  */
 'use strict'
 
-export const CHANGE_ITEM_LIST = 'CHANGE_ITEM_LIST'
 export const SIGNIN_USERNAME = 'SIGNIN_USERNAME'
 export const ADD_ORDER = 'ADD_ORDER'
 
@@ -24,9 +23,6 @@ const makeOrder = (order) => {
 }
 export const shopReducer = (state, action) => {
   switch (action.type) {
-  case CHANGE_ITEM_LIST:
-    return { ...state, jsonfile: action.payload.jsonfile, }
-
   case SIGNIN_USERNAME:
     return { ...state, username: action.payload.username, }
