+++ src/store.js
@@ -6,31 +6,24 @@
  */
 'use strict'
 
-import React, { useReducer } from 'react'
+import React, { createContext, useReducer } from 'react'
 import { cartReducer } from './reducers/cartReducer.js'
 import { shopReducer } from './reducers/shopReducer.js'
-import { POINT_INIT_VAL, JSON_INIT_VAL} from './constants'
+import { POINT_INIT_VAL } from './constants'
 
-export const ShopContext = React.createContext({})
-export const CartContext = React.createContext({})
+export const ShopContext = createContext({})
+export const CartContext = createContext({})
 
-export const ShopContextProvider = (props) => {
-  const initialState = { jsonfile: JSON_INIT_VAL, username: '' , order: []}
-  const [state, dispatch] = useReducer(shopReducer, initialState)
+export const StoreContextProvider = (props) => {
+  const [shopState, shopDispatch] = useReducer(shopReducer, {username: '' ,order: []})
+  const [cartState, cartDispatch] = useReducer(cartReducer, {cart: [], point: POINT_INIT_VAL})
 
+  // console.log('StoreContextProvider render')
   return (
-    <ShopContext.Provider value={{state,dispatch}}>
-      {props.children}
+    <ShopContext.Provider value={{state:shopState, dispatch:shopDispatch}}>
+      <CartContext.Provider value={{state:cartState, dispatch:cartDispatch}}>
+        {props.children}
+      </CartContext.Provider>
     </ShopContext.Provider>
-  )
-}
-export const CartContextProvider = (props) => {
-  const initialState = { cart: [], point: POINT_INIT_VAL }
-  const [state, dispatch] = useReducer(cartReducer, initialState)
-
-  return (
-    <CartContext.Provider value={{state,dispatch}}>
-      {props.children}
-    </CartContext.Provider>
   )
 }
